+++ libs/community/langchain_community/indexes/prompts/entity_summarization.py
