+++ libs/community/langchain_community/indexes/prompts/knowledge_triplet_extraction.py
